+++ app/services/bot_context_v2/handle_command_service.rb
@@ -6,6 +6,7 @@ module BotContextV2
       default_roll_command: 'services.bot_context_v2.commands.rolls.default',
       duality_roll_command: 'services.bot_context_v2.commands.rolls.duality',
       fate_roll_command: 'services.bot_context_v2.commands.rolls.fate',
+      cosmere_roll_command: 'services.bot_context_v2.commands.rolls.cosmere',
       check_command: 'services.bot_context_v2.commands.check'
     ]
 
@@ -14,6 +15,7 @@ module BotContextV2
       when '/roll' then default_roll_command.call(arguments: arguments)
       when '/dualityRoll' then duality_roll_command.call(arguments: arguments)
       when '/fateRoll' then fate_roll_command.call(arguments: arguments)
+      when '/plotRoll' then cosmere_roll_command.call(arguments: arguments)
       when '/check' then check_command.call(arguments: arguments, character: character)
       end
     end